change double to interger on histogram answer type

+++ model-jackson/src/main/java/io/sidecar/query/StatsHistogramAnswerMixin.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @SuppressWarnings("unused") //Used by Jackson
 abstract public class StatsHistogramAnswerMixin {
 
-    private StatsHistogramAnswerMixin(@JsonProperty("buckets") Map<String, Double> buckets) {
+    private StatsHistogramAnswerMixin(@JsonProperty("buckets") Map<String, Integer> buckets) {
 
     }
 }
