+++ model/src/test/java/io/sidecar/query/StatsHistogramAnswerTest.java
@@ -12,8 +12,8 @@ public class StatsHistogramAnswerTest {
     @Test
     public void testStatsHistogramAnswer() {
 
-        Map<String, Double> bucketsAndValues = new HashMap<>();
-        bucketsAndValues.put("0", 42.0);
+        Map<String, Integer> bucketsAndValues = new HashMap<>();
+        bucketsAndValues.put("0", 42);
 
         StatsHistogramAnswer answer = new StatsHistogramAnswer(bucketsAndValues);
